Moved mocks, deleted unused files, added todo items.
