--- Example/SuperAwesome/MoPubController.h
@@ -1,13 +1,0 @@
-//
-//  MoPubController.h
-//  SuperAwesome
-//
-//  Created by Gabriel Coman on 26/05/2017.
-//  Copyright © 2017 Gabriel Coman. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//@interface MoPubController : UIViewController
-//
-//@end
