+++ Pod/Classes/Libs/SASession/Session/SASession.m
@@ -60,6 +60,7 @@
         _bundleId = [[NSBundle mainBundle] bundleIdentifier];
         _appName = [SAUtils encodeURI:[[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleName"]];
         _device = [SAUtils getSystemSize] == size_phone ? DEVICE_PHONE : DEVICE_TABLET;
+        // TODO: Convert to use injection
         _userAgent = SADependencyContainer.shared.modules.componentModule.userAgent.name;
         _connectivityType = [SAUtils getNetworkConnectivity];
         _instl = IN_FULLSCREEN;
